requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
tavily-python
twilio
psycopg2-binary
sqlalchemy